refactor: move review-related imports from models to controllers for better organization
